requirements.txt
Flask==2.3.2
python-telegram-bot==20.8